separate audioloader module
